Input handling works, can rotate player using 'a' and 'd' keys
* Setup code that reads input through the 'input.cpp/h' module
and rotates player when pressing 'a' and 'd' keys

* Added some code comments to explain function calls in 'init_input' that
setup ncurses

* Had to replace calls to 'printf' and call ncurse's 'printw' instead.
When using ncurses, it takes control over the terminal output and so
trying to use printf in that case only results in chaos.

* Commented-out the cide that tests 'input.cpp/.h' module, might still provide
some use later on, eventually when  clean up the code I will remove it.

* Replaced printf with ansi-escape code to place cursor at top left
(printf("\033[0;0f") with call to ncurses function 'move' which does the same
thing. I also learned while I was changing this that ncurses 'printw' doesn't
handle ansi-escape codes. So to solve that I guess you have to do what I did
here, find a ncurses function that does the same thing.